LayoutTests/http/wpt/webrtc/audio-script-transform.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="video" autoplay playsInline></video>
        <script src="routines.js"></script>
        <script>
function waitForMessage(port, data)
{
    let gotMessage;
    const promise = new Promise((resolve, reject) => {
        gotMessage = resolve;
        setTimeout(() => { reject("did not get " + data) }, 5000);
    });
    port.onmessage = event => {
       if (event.data === data)
           gotMessage();
    };
    return promise;
}

let senderTransform, receiverTransform;
let stream;
promise_test(async (test) => {
    worker = new Worker("context-transform.js");
    const data = await new Promise(resolve => worker.onmessage = (event) => resolve(event.data));
    assert_equals(data, "registered");

    const localStream = await navigator.mediaDevices.getUserMedia({audio: true});

    let sender, receiver;
    senderTransform = new RTCRtpScriptTransform(worker, {name:'MockRTCRtpTransform', mediaType:'audio', side:'sender'});
    receiverTransform = new RTCRtpScriptTransform(worker, {name:'MockRTCRtpTransform', mediaType:'audio', side:'receiver'});

    promise1 = waitForMessage(senderTransform.port, "started audio sender");
    promise2 = waitForMessage(receiverTransform.port, "started audio receiver");

    stream = await new Promise((resolve, reject) => {
        createConnections((firstConnection) => {
            sender = firstConnection.addTrack(localStream.getAudioTracks()[0], localStream);
            sender.transform = senderTransform;
        }, (secondConnection) => {
            secondConnection.ontrack = (trackEvent) => {
                receiver = trackEvent.receiver;
                receiver.transform = receiverTransform;
                resolve(trackEvent.streams[0]);
            };
        });
        setTimeout(() => reject("Test timed out"), 5000);
    });

    await promise1;
    await promise2;
}, "setup and check context");

promise_test((test) => {
    video.srcObject = stream;
    return video.play();
}, "audio is playing");
        </script>
    </body>
</html>
